requirements.txt
discord
wordhoard
mysql.connector
asyncio
requests